Remove jw-reset class
The `.jw-reset` class can be safely replaced with the more useful `.jwplayer *`
selector. They have the same specificity but don't require us to remember to
add the class to every element. This also exposed a few style bleed bugs and
fixed some outstanding issues on the CSS Stylesheets Reset Test.